tests/docker/requirements-test.txt
# Python dependencies for testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-timeout>=2.1.0
requests>=2.28.0
websockets>=10.0
pyzmq>=24.0.0
psutil>=5.9.0
fastapi>=0.95.0
uvicorn>=0.20.0
httpx>=0.24.0

# Additional testing tools
pytest-mock>=3.10.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0